requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.2
flake8==7.0.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jedi==0.19.1
Mako==1.3.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
multidict==6.0.5
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pydantic==2.7.0
pydantic_core==2.18.1
pyflakes==3.2.0
Pygments==2.17.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
traitlets==5.14.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
yarl==1.9.4
